Updates to elb sync and agent startup/shutdown
