+++ BaragonAgentService/src/main/java/com/hubspot/baragon/agent/resources/RequestResource.java
@@ -12,8 +12,6 @@ import javax.ws.rs.core.Response;
 import com.google.common.base.Optional;
 import com.hubspot.baragon.agent.config.LoadBalancerConfiguration;
 import com.hubspot.baragon.agent.managers.AgentRequestManager;
-import com.hubspot.baragon.data.BaragonKnownAgentsDatastore;
-import com.hubspot.baragon.models.BaragonAgentMetadata;
 import com.hubspot.baragon.models.RequestAction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
