+++ pom.xml
@@ -59,7 +59,7 @@
     <ning.async.version>1.8.12</ning.async.version>
     <snappy.version>0.3</snappy.version>
     <sentry.version>5.0</sentry.version>
-    <horizon.version>0.0.10</horizon.version>
+    <horizon.version>0.0.13</horizon.version>
     <commons-exec.version>1.1</commons-exec.version>
     <jukito.version>1.3</jukito.version>
     <hibernate.version>5.1.1.Final</hibernate.version>
@@ -245,6 +245,12 @@
       <dependency>
         <groupId>com.hubspot</groupId>
         <artifactId>HorizonNing</artifactId>
+        <version>${horizon.version}</version>
+      </dependency>
+
+      <dependency>
+        <groupId>com.hubspot</groupId>
+        <artifactId>HorizonApache</artifactId>
         <version>${horizon.version}</version>
       </dependency>
